requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
sentence-transformers
scikit-learn
langchain
langchain-community
langchain-huggingface
chromadb
huggingface-hub
pgvector
psycopg2-binary
pypdf
openpyxl
python-docx